clean up create_secret
